pyproject.toml
[project]
name = "multi-modal-embed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.13.12",
    "numpy>=2.2.3",
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.6",
]

[tool.ruff]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "venv",
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = ["F", "E", "W", "I", "N", "UP"]
ignore = ["E501", "N"]
fixable = ["F", "E", "W", "I", "N", "UP"]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
